Change many Vector3Ds to Point3Ds in astro lib.
Positions are now all Point3D; velocities remain Vector3D. This is logically correct, though in practice it was necessary in many computations to convert points to vectors. Sometimes this was because the position coordinates also represent the vector from the origin, often another body, so that, e.g., its length is significant. Another reason for the conversion is that some general low-level routines, particularly JPLEphemeris::ComputeComponents(), are used to compute both positions and other quantities (e.g., nutation).

For this reason, I added routines to Point2 and Point3 to convert to and from vectors.