Fixed Logger for 32-bit Linux.
On a 32-bit system it appears that overloading Log(int, char * format, ...)
and Log(int, char * format, va_list) causes confusion when called with
Log(x, "%s", someCharStar). So I renamed the second variant to LogVA().
I also got a warning about converting a string literal to char * being
deprecated, so I changed a literal to a variable in a test.